+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - updating `visualization/dev-instance` to use generic parameters
 - updating `visualization/dev-instance-foxbox` to use generic parameters
 - removing `post-processing/lane-detection` due to inactivity/usage
+- made `rosbag-image-pipeline` module generic and bumped CDK version
 
 ### **Removed**
 - moved module `integration/fsx-lustre-on-eks` to [IDF-Modules repo](https://github.com/awslabs/idf-modules)